Code refactoring
- create new header file for flasher states
- create new header/source files for worker class
- avoid extra copy of QJsonArray when SocketClient object is created
